+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Add missing `__broccoliFeatures__` object
+
 # 1.0.0
 
 * Initial release